requirements.txt
pandas==2.2.3
scikit-learn==1.6.0
redis==5.2.1
aiogram==3.15.0
pillow==11.0.0